fix tests and more refactor
