+++ pkg/skaffold/deploy/helm/helm_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Skaffold Authors
+Copyright 2022 The Skaffold Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -18,438 +18,26 @@ package helm
 
 import (
 	"context"
-	"fmt"
+	"errors"
 	"io"
 	"io/ioutil"
-	"path/filepath"
-	"strings"
 	"testing"
-
-	"github.com/mitchellh/go-homedir"
-	"github.com/pkg/errors"
-	"k8s.io/client-go/tools/clientcmd/api"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/kubectl"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/label"
-	deployutil "github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/util"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/hooks"
 	ctl "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubectl"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/client"
-	kubectx "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/context"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/logger"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner/runcontext"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	schemautil "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/util"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/warnings"
 	"github.com/GoogleContainerTools/skaffold/testutil"
+	"github.com/blang/semver"
 )
 
-var testBuilds = []graph.Artifact{{
-	ImageName: "skaffold-helm",
-	Tag:       "docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184",
-}}
-
-var testBuildsFoo = []graph.Artifact{{
-	ImageName: "foo",
-	Tag:       "foo:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184",
-}}
-
-var testDeployConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-	}},
-}
-
-var testDeployNamespacedConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		Namespace: "testReleaseNamespace",
-	}},
-}
-
-var testDeployEnvTemplateNamespacedConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		Namespace: "testRelease{{.FOO}}Namespace",
-	}},
-}
-
-var testDeployConfigRemoteRepo = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		Repo: "https://charts.helm.sh/stable",
-	}},
-}
-
-var testDeployConfigTemplated = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValueTemplates: map[string]string{
-			"some.key":    "somevalue",
-			"other.key":   "{{.FOO}}",
-			"missing.key": `{{default "<MISSING>" .MISSING}}`,
-			"image.name":  "{{.IMAGE_NAME}}",
-			"image.tag":   "{{.DIGEST}}",
-			"{{.FOO}}":    "somevalue",
-		},
-	}},
-}
-
-var testDeployConfigValuesFilesTemplated = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		ValuesFiles: []string{
-			"/some/file-{{.FOO}}.yaml",
-		},
-	}},
-}
-
-var testDeployConfigVersionTemplated = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Version: "{{.VERSION}}",
-	}},
-}
-
-var testDeployConfigSetFiles = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetFiles: map[string]string{
-			"expanded": "~/file.yaml",
-			"value":    "/some/file.yaml",
-		},
-	}},
-}
-
-var testDeployRecreatePodsConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		RecreatePods: true,
-	}},
-}
-
-var testDeploySkipBuildDependenciesConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		SkipBuildDependencies: true,
-	}},
-}
-
-var testDeployHelmStyleConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		ImageStrategy: latest.HelmImageStrategy{
-			HelmImageConfig: latest.HelmImageConfig{
-				HelmConventionConfig: &latest.HelmConventionConfig{},
-			},
-		},
-	}},
-}
-
-var testDeployHelmExplicitRegistryStyleConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		ImageStrategy: latest.HelmImageStrategy{
-			HelmImageConfig: latest.HelmImageConfig{
-				HelmConventionConfig: &latest.HelmConventionConfig{
-					ExplicitRegistry: true,
-				},
-			},
-		},
-	}},
-}
-
-var testDeployConfigParameterUnmatched = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm-unmatched",
-		}},
-	},
-}
-
-var testDeployFooWithPackaged = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "foo",
-		ChartPath: "testdata/foo",
-		ArtifactOverrides: map[string]string{
-			"image": "foo",
-		},
-		Packaged: &latest.HelmPackaged{
-			Version:    "0.1.2",
-			AppVersion: "1.2.3",
-		},
-	}},
-}
-
-var testDeployWithTemplatedName = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "{{.USER}}-skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image.tag": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		}},
-	},
-}
-
-var testDeploySkipBuildDependencies = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "stable/chartmuseum",
-		ArtifactOverrides: map[string]string{
-			"image.tag": "skaffold-helm",
-		},
-		SkipBuildDependencies: true,
-	}},
-}
-
-var testDeployRemoteChart = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:        "skaffold-helm-remote",
-		RemoteChart: "stable/chartmuseum",
-		Repo:        "https://charts.helm.sh/stable",
-	}},
-}
-
-var testDeployRemoteChartVersion = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:        "skaffold-helm-remote",
-		RemoteChart: "stable/chartmuseum",
-		Version:     "1.0.0",
-		Repo:        "https://charts.helm.sh/stable",
-	}},
-}
-
-var upgradeOnChangeFalse = false
-var testDeployUpgradeOnChange = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:            "skaffold-helm-upgradeOnChange",
-		ChartPath:       "examples/test",
-		UpgradeOnChange: &upgradeOnChangeFalse,
-	}},
-}
-
-var testDeployWithoutTags = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-	}},
-}
-
-var testTwoReleases = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "other",
-		ChartPath: "examples/test",
-	}, {
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image.tag": "skaffold-helm",
-		},
-	}},
-}
-
-var createNamespaceFlag = true
-var testDeployCreateNamespaceConfig = latest.HelmDeploy{
-	Releases: []latest.HelmRelease{{
-		Name:      "skaffold-helm",
-		ChartPath: "examples/test",
-		ArtifactOverrides: map[string]string{
-			"image": "skaffold-helm",
-		},
-		Overrides: schemautil.HelmOverrides{Values: map[string]interface{}{"foo": "bar"}},
-		SetValues: map[string]string{
-			"some.key": "somevalue",
-		},
-		Namespace:       "testReleaseNamespace",
-		CreateNamespace: &createNamespaceFlag,
-	}},
-}
-
-var validDeployYaml = `
-# Source: skaffold-helm/templates/deployment.yaml
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: skaffold-helm
-  labels:
-    app: skaffold-helm
-    chart: skaffold-helm-0.1.0
-    release: skaffold-helm
-    heritage: Tiller
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: skaffold-helm
-      release: skaffold-helm
-  template:
-    metadata:
-      labels:
-        app: skaffold-helm
-        release: skaffold-helm
-    spec:
-      containers:
-        - name: skaffold-helm
-          image: gcr.io/nick-cloudbuild/skaffold-helm:f759510436c8fd6f7ffa13dd9e9d85e64bec8d2bfd12c5aa3fb9af1288eccdab
-          imagePullPolicy:
-          command: ["/bin/bash", "-c", "--" ]
-          args: ["while true; do sleep 30; done;"]
-          resources:
-            {}
-`
-
-var validServiceYaml = `
-# Source: skaffold-helm/templates/service.yaml
-apiVersion: v1
-kind: Service
-metadata:
-  name: skaffold-helm-skaffold-helm
-  labels:
-    app: skaffold-helm
-    chart: skaffold-helm-0.1.0
-    release: skaffold-helm
-    heritage: Tiller
-spec:
-  type: ClusterIP
-  ports:
-    - port: 80
-      targetPort: 80
-      protocol: TCP
-      name: nginx
-  selector:
-    app: skaffold-helm
-    release: skaffold-helm
-`
-
-var invalidDeployYaml = `REVISION: 2
-RELEASED: Tue Jun 12 15:40:18 2018
-CHART: skaffold-helm-0.1.0
-USER-SUPPLIED VALUES:
-image: gcr.io/nick-cloudbuild/skaffold-helm:f759510436c8fd6f7ffa13dd9e9d85e64bec8d2bfd12c5aa3fb9af1288eccdab
-
-COMPUTED VALUES:
-image: gcr.io/nick-cloudbuild/skaffold-helm:f759510436c8fd6f7ffa13dd9e9d85e64bec8d2bfd12c5aa3fb9af1288eccdab
-ingress:
-  annotations: null
-  enabled: false
-  hosts:
-  - chart-example.local
-  tls: null
-replicaCount: 1
-resources: {}
-service:
-  externalPort: 80
-  internalPort: 80
-  name: nginx
-  type: ClusterIP
-
-HOOKS:
-MANIFEST:
-`
-
-var (
-	// Output strings to emulate different versions of Helm
-	version20rc = `Client: &version.Version{SemVer:"v2.0.0-rc.1", GitCommit:"92be174acf51e60a33287fb7011f4571eaa5cb98", GitTreeState:"clean"}\nError: cannot connect to Tiller\n`
-	version21   = `Client: &version.Version{SemVer:"v2.15.1", GitCommit:"cf1de4f8ba70eded310918a8af3a96bfe8e7683b", GitTreeState:"clean"}\nServer: &version.Version{SemVer:"v2.16.1", GitCommit:"bbdfe5e7803a12bbdf97e94cd847859890cf4050", GitTreeState:"clean"}\n`
-	version30b  = `version.BuildInfo{Version:"v3.0.0-beta.3", GitCommit:"5cb923eecbe80d1ad76399aee234717c11931d9a", GitTreeState:"clean", GoVersion:"go1.12.9"}`
-	version30   = `version.BuildInfo{Version:"v3.0.0", GitCommit:"e29ce2a54e96cd02ccfce88bee4f58bb6e2a28b6", GitTreeState:"clean", GoVersion:"go1.13.4"}`
-	version31   = `version.BuildInfo{Version:"v3.1.1", GitCommit:"afe70585407b420d0097d07b21c47dc511525ac8", GitTreeState:"clean", GoVersion:"go1.13.8"}`
-	version32   = `version.BuildInfo{Version:"v3.2.0", GitCommit:"e11b7ce3b12db2941e90399e874513fbd24bcb71", GitTreeState:"clean", GoVersion:"go1.14"}`
-	version35   = `version.BuildInfo{Version:"3.5.2", GitCommit:"c4e74854886b2efe3321e185578e6db9be0a6e29", GitTreeState:"clean", GoVersion:"go1.14.15"}`
-)
-
-func TestBinVer(t *testing.T) {
+func TestBin(t *testing.T) {
 	tests := []struct {
 		description string
 		helmVersion string
@@ -471,636 +59,6 @@ func TestBinVer(t *testing.T) {
 			ver, err := binVer(context.Background())
 
 			t.CheckErrorAndDeepEqual(test.shouldErr, err, test.expected, ver.String())
-		})
-	}
-}
-
-func TestNewDeployer(t *testing.T) {
-	tests := []struct {
-		description string
-		helmVersion string
-		shouldErr   bool
-	}{
-		{"Helm 2.0RC1", version20rc, true},
-		{"Helm 2.15.1", version21, true},
-		{"Helm 3.0.0-beta.0", version30b, false},
-		{"Helm 3.0", version30, false},
-		{"Helm 3.1.1", version31, false},
-		{"helm3 unparseable version", "gobbledygook", true},
-	}
-
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", test.helmVersion))
-
-			_, err := NewDeployer(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &testDeployConfig)
-			t.CheckError(test.shouldErr, err)
-		})
-	}
-}
-
-func TestHelmDeploy(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "TestHelmDeploy")
-	if err != nil {
-		t.Fatalf("tempdir: %v", err)
-	}
-	home, err := homedir.Dir()
-	if err != nil {
-		t.Fatalf("Cannot get homedir: %v", err)
-	}
-
-	tests := []struct {
-		description        string
-		commands           util.Command
-		env                []string
-		helm               latest.HelmDeploy
-		namespace          string
-		configure          func(*Deployer3)
-		builds             []graph.Artifact
-		force              bool
-		shouldErr          bool
-		expectedWarnings   []string
-		expectedNamespaces []string
-	}{
-		{
-			description: "helm3.0beta deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30b).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.0beta namespaced context deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30b).
-				AndRun("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			namespace:          kubectl.TestNamespace,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testNamespace"},
-		},
-		{
-			description: "helm3.0 deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.0 namespaced deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30).
-				AndRun("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testReleaseNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployNamespacedConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseNamespace"},
-		},
-		{
-			description: "helm3.0 namespaced (with env template) deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30).
-				AndRun("helm --kube-context kubecontext get all --namespace testReleaseFOOBARNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testReleaseFOOBARNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseFOOBARNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployEnvTemplateNamespacedConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseFOOBARNamespace"},
-		},
-		{
-			description: "helm3.0 namespaced context deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30).
-				AndRun("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			namespace:          kubectl.TestNamespace,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testNamespace"},
-		},
-		{
-			description: "helm3.0 namespaced context deploy success overrides release namespaces",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30).
-				AndRun("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployNamespacedConfig,
-			namespace:          kubectl.TestNamespace,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testNamespace"},
-		},
-		{
-			description: "helm3.1 deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.1 namespaced deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testReleaseNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployNamespacedConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseNamespace"},
-		},
-		{
-			description: "helm3.1 namespaced deploy (with env template) success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all --namespace testReleaseFOOBARNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testReleaseFOOBARNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseFOOBARNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployEnvTemplateNamespacedConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseFOOBARNamespace"},
-		},
-		{
-			description: "helm3.1 deploy with repo success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --repo https://charts.helm.sh/stable --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfigRemoteRepo,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.1 namespaced context deploy success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			namespace:          kubectl.TestNamespace,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testNamespace"},
-		},
-		{
-			description: "helm3.1 namespaced context deploy success overrides release namespaces",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployNamespacedConfig,
-			namespace:          kubectl.TestNamespace,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testNamespace"},
-		},
-		{
-			description: "deploy success with recreatePods",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm --recreate-pods examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployRecreatePodsConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy success with skipBuildDependencies",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeploySkipBuildDependenciesConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy should error for unmatched parameter",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfigParameterUnmatched,
-			builds:             testBuilds,
-			shouldErr:          true,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy success remote chart with skipBuildDependencies",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm stable/chartmuseum --set-string image.tag=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeploySkipBuildDependencies,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy success when `upgradeOnChange: false` and does not upgrade",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm-upgradeOnChange --kubeconfig kubeconfig"),
-			helm:               testDeployUpgradeOnChange,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy remote chart",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm-remote --kubeconfig kubeconfig", fmt.Errorf("Error: release: not found")).
-				AndRun("helm --kube-context kubecontext install skaffold-helm-remote stable/chartmuseum --repo https://charts.helm.sh/stable --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm-remote --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployRemoteChart,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy remote chart with version",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm-remote --kubeconfig kubeconfig", fmt.Errorf("Error: release: not found")).
-				AndRun("helm --kube-context kubecontext install skaffold-helm-remote --version 1.0.0 stable/chartmuseum --repo https://charts.helm.sh/stable --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm-remote --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployRemoteChartVersion,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy error with remote chart",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm-remote --kubeconfig kubeconfig", fmt.Errorf("Error: release: not found")).
-				AndRunErr("helm --kube-context kubecontext install skaffold-helm-remote stable/chartmuseum --repo https://charts.helm.sh/stable --kubeconfig kubeconfig", fmt.Errorf("building helm dependencies")),
-			helm:               testDeployRemoteChart,
-			shouldErr:          true,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "get failure should install not upgrade",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext install skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3 get failure should install not upgrade",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext install skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "get failure should install not upgrade with helm image strategy",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext install skaffold-helm examples/test --set-string image.repository=docker.io:5000/skaffold-helm,image.tag=3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployHelmStyleConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm image strategy with explicit registry should set the Helm registry value",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext install skaffold-helm examples/test --set-string image.registry=docker.io:5000,image.repository=skaffold-helm,image.tag=3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployHelmExplicitRegistryStyleConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "get success should upgrade by force, not install",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm --force examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			force:              true,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "get success should upgrade without force, not install",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy error",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRunErr("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig", fmt.Errorf("unexpected error")).
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			shouldErr:          true,
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "dep build error",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRunErr("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig", fmt.Errorf("unexpected error")).
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			shouldErr:          true,
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm 3.0 beta should package chart and deploy",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version30b).
-				AndRun("helm --kube-context kubecontext get all foo --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build testdata/foo --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext package testdata/foo --destination "+tmpDir+" --version 0.1.2 --app-version 1.2.3 --kubeconfig kubeconfig", fmt.Sprintf("Packaged to %s", filepath.Join(tmpDir, "foo-0.1.2.tgz"))).
-				AndRun("helm --kube-context kubecontext upgrade foo "+filepath.Join(tmpDir, "foo-0.1.2.tgz")+" --set-string image=foo:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all foo --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			shouldErr:          false,
-			helm:               testDeployFooWithPackaged,
-			builds:             testBuildsFoo,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm 3.1 should package chart and deploy",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all foo --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build testdata/foo --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext package testdata/foo --destination "+tmpDir+" --version 0.1.2 --app-version 1.2.3 --kubeconfig kubeconfig", fmt.Sprintf("Packaged to %s", filepath.Join(tmpDir, "foo-0.1.2.tgz"))).
-				AndRun("helm --kube-context kubecontext upgrade foo "+filepath.Join(tmpDir, "foo-0.1.2.tgz")+" --set-string image=foo:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all foo --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			shouldErr:          false,
-			helm:               testDeployFooWithPackaged,
-			builds:             testBuildsFoo,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "should fail to deploy when packaging fails",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all foo --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build testdata/foo --kubeconfig kubeconfig").
-				AndRunErr("helm --kube-context kubecontext package testdata/foo --destination "+tmpDir+" --version 0.1.2 --app-version 1.2.3 --kubeconfig kubeconfig", fmt.Errorf("packaging failed")),
-			shouldErr:          true,
-			helm:               testDeployFooWithPackaged,
-			builds:             testBuildsFoo,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description:        "deploy and get missing templated release name should fail",
-			commands:           testutil.CmdRunWithOutput("helm version --client", version31),
-			helm:               testDeployWithTemplatedName,
-			builds:             testBuilds,
-			shouldErr:          true,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy and get templated release name",
-			env:         []string{"USER=user"},
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all user-skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade user-skaffold-helm examples/test --set-string image.tag=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all user-skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployWithTemplatedName,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy with templated values",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set image.name=skaffold-helm --set image.tag=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set missing.key=<MISSING> --set other.key=FOOBAR --set some.key=somevalue --set FOOBAR=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfigTemplated,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy with valuesFiles templated",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 -f /some/file-FOOBAR.yaml -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfigValuesFilesTemplated,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy with templated version",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm --version 1.0 examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			env:                []string{"VERSION=1.0"},
-			helm:               testDeployConfigVersionTemplated,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy with setFiles",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun(fmt.Sprintf("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set-file expanded=%s --set-file value=/some/file.yaml -f skaffold-overrides.yaml --kubeconfig kubeconfig", strings.ReplaceAll(filepath.Join(home, "file.yaml"), "\\", "\\\\"))).
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfigSetFiles,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "deploy without actual tags",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:   testDeployWithoutTags,
-			builds: testBuilds,
-			expectedWarnings: []string{
-				"See helm documentation on how to replace image names with their actual tags: https://skaffold.dev/docs/pipeline-stages/deployers/helm/#image-configuration",
-				"image [docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184] is not used.",
-			},
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "first release without tag, second with tag",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all other --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade other examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext get all other --template {{.Release.Manifest}} --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --set-string image.tag=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testTwoReleases,
-			builds:             testBuilds,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description:        "debug for helm3.0 failure",
-			commands:           testutil.CmdRunWithOutput("helm version --client", version30),
-			shouldErr:          true,
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			configure:          func(deployer *Deployer3) { deployer.enableDebug = true },
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "debug for helm3.1 success",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRun("helm --kube-context kubecontext get all skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRunEnv("helm --kube-context kubecontext upgrade skaffold-helm --post-renderer SKAFFOLD-BINARY examples/test --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig",
-					[]string{"SKAFFOLD_FILENAME=test.yaml"}).
-				AndRunWithOutput("helm --kube-context kubecontext get all skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployConfig,
-			builds:             testBuilds,
-			configure:          func(deployer *Deployer3) { deployer.enableDebug = true },
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.1 should fail to deploy with createNamespace option",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version31).
-				AndRunErr("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig"),
-			helm:               testDeployCreateNamespaceConfig,
-			builds:             testBuilds,
-			shouldErr:          true,
-			expectedNamespaces: []string{""},
-		},
-		{
-			description: "helm3.2 get failure should install with createNamespace not upgrade",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version32).
-				AndRunErr("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --kubeconfig kubeconfig", fmt.Errorf("not found")).
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext install skaffold-helm examples/test --namespace testReleaseNamespace --create-namespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployCreateNamespaceConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseNamespace"},
-		},
-		{
-			description: "helm3.2 namespaced deploy success without createNamespace",
-			commands: testutil.
-				CmdRunWithOutput("helm version --client", version32).
-				AndRun("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext dep build examples/test --kubeconfig kubeconfig").
-				AndRun("helm --kube-context kubecontext upgrade skaffold-helm examples/test --namespace testReleaseNamespace --set-string image=docker.io:5000/skaffold-helm:3605e7bc17cf46e53f4d81c4cbc24e5b4c495184 --set some.key=somevalue -f skaffold-overrides.yaml --kubeconfig kubeconfig").
-				AndRunWithOutput("helm --kube-context kubecontext get all --namespace testReleaseNamespace skaffold-helm --template {{.Release.Manifest}} --kubeconfig kubeconfig", validDeployYaml),
-			helm:               testDeployCreateNamespaceConfig,
-			builds:             testBuilds,
-			expectedNamespaces: []string{"testReleaseNamespace"},
-		},
-	}
-
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			t.Override(&client.Client, deployutil.MockK8sClient)
-			fakeWarner := &warnings.Collect{}
-			env := test.env
-			if env == nil {
-				env = []string{"FOO=FOOBAR"}
-			}
-			t.Override(&warnings.Printf, fakeWarner.Warnf)
-			t.Override(&util.OSEnviron, func() []string { return env })
-			t.Override(&util.DefaultExecCommand, test.commands)
-			t.Override(&osExecutable, func() (string, error) { return "SKAFFOLD-BINARY", nil })
-			t.Override(&kubectx.CurrentConfig, func() (api.Config, error) {
-				return api.Config{CurrentContext: ""}, nil
-			})
-
-			deployer, err := NewDeployer(context.Background(), &helmConfig{
-				namespace:  test.namespace,
-				force:      test.force,
-				configFile: "test.yaml",
-			}, &label.DefaultLabeller{}, &test.helm)
-			t.RequireNoError(err)
-
-			if test.configure != nil {
-				test.configure(deployer)
-			}
-			deployer.pkgTmpDir = tmpDir
-			// Deploy returns nil unless `helm get all <release>` is set up to return actual release info
-			err = deployer.Deploy(context.Background(), ioutil.Discard, test.builds)
-			t.CheckError(test.shouldErr, err)
-			t.CheckDeepEqual(test.expectedWarnings, fakeWarner.Warnings)
-			t.CheckErrorAndDeepEqual(test.shouldErr, err, test.expectedNamespaces, *deployer.namespaces)
 		})
 	}
 }
@@ -1211,6 +169,157 @@ func TestParseHelmRelease(t *testing.T) {
 			_, err := parseRuntimeObject(kubectl.TestNamespace, test.yaml)
 
 			t.CheckError(test.shouldErr, err)
+		})
+	}
+}
+
+func TestWriteBuildArtifacts(t *testing.T) {
+	tests := []struct {
+		description string
+		builds      []graph.Artifact
+		result      string
+	}{
+		{
+			description: "nil",
+			builds:      nil,
+			result:      `{"builds":null}`,
+		},
+		{
+			description: "empty",
+			builds:      []graph.Artifact{},
+			result:      `{"builds":[]}`,
+		},
+		{
+			description: "multiple images with tags",
+			builds:      []graph.Artifact{{ImageName: "name", Tag: "name:tag"}, {ImageName: "name2", Tag: "name2:tag"}},
+			result:      `{"builds":[{"imageName":"name","tag":"name:tag"},{"imageName":"name2","tag":"name2:tag"}]}`,
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			file, cleanup, err := writeBuildArtifacts(test.builds)
+			t.CheckError(false, err)
+			if content, err := ioutil.ReadFile(file); err != nil {
+				t.Errorf("error reading file %q: %v", file, err)
+			} else {
+				t.CheckDeepEqual(test.result, string(content))
+			}
+			cleanup()
+		})
+	}
+}
+
+func TestGenerateSkaffoldDebugFilter(t *testing.T) {
+	tests := []struct {
+		description string
+		buildFile   string
+		result      []string
+	}{
+		{
+			description: "empty buildfile is skipped",
+			buildFile:   "",
+			result:      []string{"filter", "--debugging", "--kube-context", "kubecontext", "--kubeconfig", "kubeconfig"},
+		},
+		{
+			description: "buildfile is added",
+			buildFile:   "buildfile",
+			result:      []string{"filter", "--debugging", "--kube-context", "kubecontext", "--build-artifacts", "buildfile", "--kubeconfig", "kubeconfig"},
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
+			h, err := NewDeployer(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &testDeployConfig)
+			h3 := h.(*Deployer31)
+			t.RequireNoError(err)
+			result := h3.generateSkaffoldDebugFilter(test.buildFile)
+			t.CheckDeepEqual(test.result, result)
+		})
+	}
+}
+
+func TestHelmHooks(t *testing.T) {
+	tests := []struct {
+		description string
+		runner      hooks.Runner
+		shouldErr   bool
+	}{
+		{
+			description: "hooks run successfully",
+			runner: hooks.MockRunner{
+				PreHooks: func(context.Context, io.Writer) error {
+					return nil
+				},
+				PostHooks: func(context.Context, io.Writer) error {
+					return nil
+				},
+			},
+		},
+		{
+			description: "hooks fails",
+			runner: hooks.MockRunner{
+				PreHooks: func(context.Context, io.Writer) error {
+					return errors.New("failed to execute hooks")
+				},
+				PostHooks: func(context.Context, io.Writer) error {
+					return errors.New("failed to execute hooks")
+				},
+			},
+			shouldErr: true,
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
+			t.Override(&hooks.NewDeployRunner, func(*ctl.CLI, latest.DeployHooks, *[]string, logger.Formatter, hooks.DeployEnvOpts) hooks.Runner {
+				return test.runner
+			})
+			v, err := semver.New("3.0.0")
+			t.RequireNoError(err)
+			k, err := NewDeployer30(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &testDeployConfig, *v)
+			t.RequireNoError(err)
+			err = k.PreDeployHooks(context.Background(), ioutil.Discard)
+			t.CheckError(test.shouldErr, err)
+			err = k.PostDeployHooks(context.Background(), ioutil.Discard)
+			t.CheckError(test.shouldErr, err)
+		})
+	}
+}
+
+func TestHasRunnableHooks(t *testing.T) {
+	tests := []struct {
+		description string
+		cfg         latest.HelmDeploy
+		expected    bool
+	}{
+		{
+			description: "no hooks defined",
+			cfg:         latest.HelmDeploy{},
+		},
+		{
+			description: "has pre-deploy hook defined",
+			cfg: latest.HelmDeploy{
+				LifecycleHooks: latest.DeployHooks{PreHooks: []latest.DeployHookItem{{}}},
+			},
+			expected: true,
+		},
+		{
+			description: "has post-deploy hook defined",
+			cfg: latest.HelmDeploy{
+				LifecycleHooks: latest.DeployHooks{PostHooks: []latest.DeployHookItem{{}}},
+			},
+			expected: true,
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
+			v, err := semver.New("3.0.0")
+			t.RequireNoError(err)
+			k, err := NewDeployer30(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &test.cfg, *v)
+			t.RequireNoError(err)
+			actual := k.HasRunnableHooks()
+			t.CheckDeepEqual(test.expected, actual)
 		})
 	}
 }
@@ -1542,6 +651,7 @@ func TestHelmRender(t *testing.T) {
 
 			t.Override(&util.OSEnviron, func() []string { return append([]string{"FOO=FOOBAR"}, test.env...) })
 			t.Override(&util.DefaultExecCommand, test.commands)
+
 			deployer, err := NewDeployer(context.Background(), &helmConfig{
 				namespace: test.namespace,
 			}, &label.DefaultLabeller{}, &test.helm)
@@ -1553,167 +663,6 @@ func TestHelmRender(t *testing.T) {
 				dat, _ := ioutil.ReadFile(file)
 				t.CheckDeepEqual(string(dat), test.expected)
 			}
-		})
-	}
-}
-
-func TestWriteBuildArtifacts(t *testing.T) {
-	tests := []struct {
-		description string
-		builds      []graph.Artifact
-		result      string
-	}{
-		{
-			description: "nil",
-			builds:      nil,
-			result:      `{"builds":null}`,
-		},
-		{
-			description: "empty",
-			builds:      []graph.Artifact{},
-			result:      `{"builds":[]}`,
-		},
-		{
-			description: "multiple images with tags",
-			builds:      []graph.Artifact{{ImageName: "name", Tag: "name:tag"}, {ImageName: "name2", Tag: "name2:tag"}},
-			result:      `{"builds":[{"imageName":"name","tag":"name:tag"},{"imageName":"name2","tag":"name2:tag"}]}`,
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			file, cleanup, err := writeBuildArtifacts(test.builds)
-			t.CheckError(false, err)
-			if content, err := ioutil.ReadFile(file); err != nil {
-				t.Errorf("error reading file %q: %v", file, err)
-			} else {
-				t.CheckDeepEqual(test.result, string(content))
-			}
-			cleanup()
-		})
-	}
-}
-
-func TestGenerateSkaffoldDebugFilter(t *testing.T) {
-	tests := []struct {
-		description string
-		buildFile   string
-		result      []string
-	}{
-		{
-			description: "empty buildfile is skipped",
-			buildFile:   "",
-			result:      []string{"filter", "--debugging", "--kube-context", "kubecontext", "--kubeconfig", "kubeconfig"},
-		},
-		{
-			description: "buildfile is added",
-			buildFile:   "buildfile",
-			result:      []string{"filter", "--debugging", "--kube-context", "kubecontext", "--build-artifacts", "buildfile", "--kubeconfig", "kubeconfig"},
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
-			h, err := NewDeployer(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &testDeployConfig)
-			t.RequireNoError(err)
-			result := h.generateSkaffoldDebugFilter(test.buildFile)
-			t.CheckDeepEqual(test.result, result)
-		})
-	}
-}
-
-func TestHelmHooks(t *testing.T) {
-	tests := []struct {
-		description string
-		runner      hooks.Runner
-		shouldErr   bool
-	}{
-		{
-			description: "hooks run successfully",
-			runner: hooks.MockRunner{
-				PreHooks: func(context.Context, io.Writer) error {
-					return nil
-				},
-				PostHooks: func(context.Context, io.Writer) error {
-					return nil
-				},
-			},
-		},
-		{
-			description: "hooks fails",
-			runner: hooks.MockRunner{
-				PreHooks: func(context.Context, io.Writer) error {
-					return errors.New("failed to execute hooks")
-				},
-				PostHooks: func(context.Context, io.Writer) error {
-					return errors.New("failed to execute hooks")
-				},
-			},
-			shouldErr: true,
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
-			t.Override(&hooks.NewDeployRunner, func(*ctl.CLI, latest.DeployHooks, *[]string, logger.Formatter, hooks.DeployEnvOpts) hooks.Runner {
-				return test.runner
-			})
-
-			k, err := NewDeployer(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &testDeployConfig)
-			t.RequireNoError(err)
-			err = k.PreDeployHooks(context.Background(), ioutil.Discard)
-			t.CheckError(test.shouldErr, err)
-			err = k.PostDeployHooks(context.Background(), ioutil.Discard)
-			t.CheckError(test.shouldErr, err)
-		})
-	}
-}
-
-type helmConfig struct {
-	runcontext.RunContext // Embedded to provide the default values.
-	namespace             string
-	force                 bool
-	configFile            string
-}
-
-func (c *helmConfig) ForceDeploy() bool                                   { return c.force }
-func (c *helmConfig) GetKubeConfig() string                               { return kubectl.TestKubeConfig }
-func (c *helmConfig) GetKubeContext() string                              { return kubectl.TestKubeContext }
-func (c *helmConfig) GetKubeNamespace() string                            { return c.namespace }
-func (c *helmConfig) ConfigurationFile() string                           { return c.configFile }
-func (c *helmConfig) PortForwardResources() []*latest.PortForwardResource { return nil }
-
-func TestHasRunnableHooks(t *testing.T) {
-	tests := []struct {
-		description string
-		cfg         latest.HelmDeploy
-		expected    bool
-	}{
-		{
-			description: "no hooks defined",
-			cfg:         latest.HelmDeploy{},
-		},
-		{
-			description: "has pre-deploy hook defined",
-			cfg: latest.HelmDeploy{
-				LifecycleHooks: latest.DeployHooks{PreHooks: []latest.DeployHookItem{{}}},
-			},
-			expected: true,
-		},
-		{
-			description: "has post-deploy hook defined",
-			cfg: latest.HelmDeploy{
-				LifecycleHooks: latest.DeployHooks{PostHooks: []latest.DeployHookItem{{}}},
-			},
-			expected: true,
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			t.Override(&util.DefaultExecCommand, testutil.CmdRunWithOutput("helm version --client", version31))
-			k, err := NewDeployer(context.Background(), &helmConfig{}, &label.DefaultLabeller{}, &test.cfg)
-			t.RequireNoError(err)
-			actual := k.HasRunnableHooks()
-			t.CheckDeepEqual(test.expected, actual)
 		})
 	}
 }
